move Share[Data/Position]Chip to chips module
