+++ src/modules/chips/components/assistChips/ShareDataChip.js
@@ -5,7 +5,7 @@ import { html } from 'lit-html';
 import { QueryParameters } from '../../../../domain/queryParameters';
 import { $injector } from '../../../../injection/index';
 import { openModal } from '../../../../store/modal/modal.action';
-import { AbstractAssistChip } from '../../../chips/components/assistChips/AbstractAssistChip';
+import { AbstractAssistChip } from './AbstractAssistChip';
 import shareIcon from './assets/share.svg';
 
 const Update = 'update';
