+++ src/modules/chips/components/assistChips/index.js
@@ -2,6 +2,8 @@ import { ElevationProfileChip } from './ElevationProfileChip';
 import { ExportVectorDataChip } from './ExportVectorDataChip';
 import { MapFeedbackChip } from './MapFeedbackChip';
 import { RoutingChip } from './RoutingChip';
+import { ShareDataChip } from './ShareDataChip';
+import { SharePositionChip } from './SharePositionChip';
 
 if (!window.customElements.get(ElevationProfileChip.tag)) {
 	window.customElements.define(ElevationProfileChip.tag, ElevationProfileChip);
@@ -17,4 +19,12 @@ if (!window.customElements.get(MapFeedbackChip.tag)) {
 
 if (!window.customElements.get(RoutingChip.tag)) {
 	window.customElements.define(RoutingChip.tag, RoutingChip);
+}
+
+if (!window.customElements.get(ShareDataChip.tag)) {
+	window.customElements.define(ShareDataChip.tag, ShareDataChip);
+}
+
+if (!window.customElements.get(SharePositionChip.tag)) {
+	window.customElements.define(SharePositionChip.tag, SharePositionChip);
 }
